Change FluentApi and models

+++ DAL/Models/Person.cs
@@ -14,13 +14,13 @@ namespace Dal
             Trainee = new HashSet<Trainee>();
         }
 
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Login { get; set; }
         public string Password { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
-        public long SexStatusId { get; set; }
+        public int SexStatusId { get; set; }
         public string Phone { get; set; }
         public bool IsDeleted { get; set; }
         public bool IsBanned { get; set; }
