+++ DAL/Models/Speciality.cs
@@ -5,8 +5,8 @@ namespace Dal
 {
     public partial class Speciality
     {
-        public long Id { get; set; }
-        public long CoachId { get; set; }
+        public int Id { get; set; }
+        public int CoachId { get; set; }
         public string Speciality1 { get; set; }
         public string Experience { get; set; }
         public int Rating { get; set; }
