+++ DAL/Models/TraineeAchivements.cs
@@ -5,9 +5,9 @@ namespace Dal
 {
     public partial class TraineeAchivements
     {
-        public long Id { get; set; }
-        public long TraineeId { get; set; }
-        public long AchivementId { get; set; }
+        public int Id { get; set; }
+        public int TraineeId { get; set; }
+        public int AchivementId { get; set; }
 
         public virtual Achivement Achivement { get; set; }
         public virtual Trainee Trainee { get; set; }
